apps/agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10, <3.15"

dependencies = [
    "livekit-agents[mcp,openai,mistralai,silero,turn-detector]==1.5.5",
    "livekit-plugins-ai-coustics~=0.2",
    # Phase-05 type_b: Mistral Agents SDK drives its own MCP tool loop via
    # mistralai.extra.mcp.stdio.MCPClientSTDIO. Kept explicit (not just as
    # transitive of livekit-plugins-mistralai) so a future LiveKit bump
    # cannot silently drop the EU-voice path.
    # `griffe` is an undeclared peer dep of mistralai.extra.run.tools in
    # 2.4.0 (import-time requirement for RunContext) — pinning it here
    # until upstream fixes the wheel metadata.
    "mistralai>=2.4.0,<3.0.0",
    "mcp>=1.0.0",
    "griffe>=0.49",
    "python-dotenv",
    "python-json-logger~=2.0.0",
    # Optional local TTS provider. Activated via TTS_PROVIDER=piper; the
    # default Voxtral path ignores it entirely. Keeping the dep here
    # (mirrored in requirements.txt) keeps Docker build and uv dev in sync.
    "piper-tts>=1.3.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
